Update structure to use single entry point for code
Use index.js as the launchpad for all days. This means that we each day
doesn't need to require the loader and configure the env variables, and
can just contain the code needed for the puzzle.

This also stops jest tests from outputting the `console.log` output for
some strange reason.